==> notebooks/1_gct_stories/1_generate/00_metadata_voxels.ipynb <==
%load_ext autoreload
%autoreload 2
import pandas as pd
story_generate = __import__('01_generate_story')
from sasc.modules.fmri_module import convert_module_num_to_voxel_num
subject = 'UTS02'
setting = 'polysemantic'
rows, _, _ = story_generate.get_rows_and_prompts_default(subject=subject,
    setting=setting,
    seed=1,
    n_examples_per_prompt_to_consider=35,
    n_examples_per_prompt=3,
    version='v5_noun')
rows['top_ngrams_module_correct'] = rows['top_ngrams_module_correct'].apply(lambda x: x[:5])
rows["voxel_num"] = rows.apply(
    lambda row: convert_module_num_to_voxel_num(row["module_num"], row["subject"]),
    axis=1,
)

# with pd.option_context('display.max_colwidth', None):
    # display(rows[['expl', 'roi_func', 'top_ngrams_module_correct']].style.hide(axis='index'))
with pd.option_context('display.max_colwidth', None):
    display(rows[['voxel_num', 'expl', 'roi_func', 'top_ngrams_module_correct']].sort_values(by='voxel_num').style.hide(axis='index'))
